WarcReader: Try to recover when the record trailer is truncated
This also adds a new mechanism for logging a warning instead of
throwing an exception so the validate tool can still report it as a
validation error.

We encountered a large number of WARCs in the wild that are missing the
trailer on warcinfo records and this will at least let us index them.

Closes #29