LenientHttpParser: Accept headers missing terminating linefeeds
We've seen in the wild some Heritrix WARC files where a redirect HTTP
response is missing the terminating CRLF at the end of the headers.
Presumably some servers send it this way when there's no message body.

To support this the parser now needs to know when we hit the end of
input. This is indicated by calling parse(null).