Keep the original HTTP header bytes for serializeHeader() when parsing
We're already making a copy of the header bytes to pushback anyway so
let's also stash them in the HttpResponse and HttpRequest objects. This
means serializeHeader() can return the original bytes making it easy to
copy the header without reformatting it.

This means that ExtractTool now preserves the original header
formatting and this can also be used in future by DedupeTool.