Add lenient mode to chunked encoding parser
Some tools such as Browsertrix will save WARC recorsd with chunked
encoding decoded but will leave the "Transfer-Encoding: chunked" HTTP
header present. This makes it difficult to always correctly decode the
body accurately since the content might start with something that looks
like a chunked encoding header. But if the body starts with something
that's definitely not a valid chunk header we can stop trying to decode
the chunked encoding. This will work a lot of the time.