CdxFormat: Fix singleton mutation bug by adding builder
setDigestUnchanged() was mutating the format singletons (e.g. CdxFormat.CDX11) which was likely to cause unexpected behaviour when used in a multithreaded on longer-running process. Since we use the builder pattern everywhere else in jwarc it probably makes sense to use it for constructing a CdxFormat with extra options too.

#74