Cdxtool output refactoring (#77)
Refactoring the WARC-record iteration and output generation to a new class. The class can be given a PrintWriter. The CDX-tool will now call the class with System.out as PrintWriter, so everything is working as before.

The refactoring will make it possible for java code make the same cdx-output as was printed to System.out before with the CDX-tool.