GunzipChannel: Ensure gzip trailer is fully consumed at end of record
If the buffer aligns just right we sometimes returned after reading the
last compressed byte but without reading the gzip trailer. This would
cause position of the next to be calculated incorrectly as it point at
the trailer of the previous gzip member rather than the start of the
next member's header.